+++ lesson_7/src/main/java/ru/khorolskii/lesson_7/lesson_7/controllers/CartController.java
@@ -2,14 +2,13 @@ package ru.khorolskii.lesson_7.lesson_7.controllers;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
 import ru.khorolskii.lesson_7.lesson_7.converter.ProductConverter;
-import ru.khorolskii.lesson_7.lesson_7.dto.ProductDto;
 import ru.khorolskii.lesson_7.lesson_7.dto.ProductToCartDto;
 import ru.khorolskii.lesson_7.lesson_7.entities.Product;
 import ru.khorolskii.lesson_7.lesson_7.exceptions.ResourceNotFoundExceptions;
 import ru.khorolskii.lesson_7.lesson_7.repositories.ProductRepository;
+import ru.khorolskii.lesson_7.lesson_7.entities.Cart;
 import ru.khorolskii.lesson_7.lesson_7.services.CartService;
 
 import java.util.List;
@@ -29,16 +28,21 @@ public class CartController {
         log.debug("id продукта, который пришел с фронта: " + productId);
         Product product = productRepository.findById(productId).orElseThrow(() -> new ResourceNotFoundExceptions("Product not found, id: " + productId));
         log.debug("отправка DTO в корзину: " + product.toString());
-        return cartService.add(product);
+        return cartService.addProduct(product);
     }
 
     @GetMapping("/remove")
     public void removeProduct(@RequestParam Long productId){
-        cartService.remove(productId);
+        cartService.removeProduct(productId);
     }
 
     @GetMapping("/show")
     public List <ProductToCartDto> showCart(){
        return cartService.showCart();
+    }
+
+    @GetMapping("/clear")
+    public void clear(){
+        cartService.clear();
     }
 }
